Partial new config

+++ JeMPI_Apps/JeMPI_LibShared/src/main/java/org/jembi/jempi/shared/config/input/MatchMetaData.java
@@ -7,7 +7,7 @@ import java.util.List;
 @JsonInclude(JsonInclude.Include.NON_NULL)
    public record MatchMetaData(
          String comparison,
-         List<Double> comparisonLevels,
-         Double m,
-         Double u) {
+         List<Float> comparisonLevels,
+         Float m,
+         Float u) {
    }
